third_party/blink/web_tests/platform/win/fast/forms/select-popup/popup-menu-mouse-operations-expected.txt

PASS picker._selectElement.value is "1"
PASS menuElement.value is "bar"
PASS picker._selectElement.value is "0"
PASS menuElement.value is "bar"
PASS picker._selectElement.value is "1"
PASS menuElement.value is "bar"
PASS picker._selectElement.value is "1"
PASS menuElement.value is "bar"
PASS internals.pagePopupWindow is not null
PASS picker._selectElement.value is "0"
PASS menuElement.value is "bar"
PASS clickEventCounter is 0
PASS mouseupEventCounter is 0
PASS internals.pagePopupWindow is null
PASS menuElement.value is "baz"
PASS clickEventCounter is 1
PASS mouseupEventCounter is 1
PASS picker._selectElement.value is "2"
PASS internals.selectMenuListText(menu) is "baz"
PASS picker._selectElement.value is "1"
PASS menu.value is "baz"
PASS internals.selectMenuListText(menu) is "bar"
PASS internals.pagePopupWindow is null
PASS menu.value is "bar"
PASS internals.selectMenuListText(menu) is "bar"
PASS picker._selectElement.selectedOptions[0].label is "option20"
PASS successfullyParsed is true

TEST COMPLETE

foo
bar
baz
 
option1
option2
option3
option4
option5
option6
option7
option8
option9
option10
option11
option12
option13
option14
option15
option16
option17
option18
option19
option20
option21
